format the file
