+++ operator/pkg/controllers/hubofhubs/conditions.go
@@ -3,12 +3,18 @@ package hubofhubs
 import (
 	"context"
 	"fmt"
+	"time"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	hubofhubsv1alpha1 "github.com/stolostron/hub-of-hubs/operator/apis/hubofhubs/v1alpha1"
-	"github.com/stolostron/hub-of-hubs/operator/pkg/utils"
 )
+
+type ICondition interface {
+	GetConditions() []metav1.Condition
+	SetConditions([]metav1.Condition)
+}
 
 const (
 	CONDITION_STATUS_TRUE    = "True"
@@ -16,23 +22,22 @@ const (
 	CONDITION_STATUS_UNKNOWN = "Unknown"
 )
 
-// NOTE: the status of RESOURCE_FOUND can only be True; otherwise there is no condition
-const CONDITION_TYPE_RESOURCE_FOUND = "ResourceFound"
-
+// NOTE: the status of ResourceFound can only be True; otherwise there is no condition
 const (
+	CONDITION_TYPE_RESOURCE_FOUND    = "ResourceFound"
 	CONDITION_REASON_RESOURCE_FOUND  = "ResourceFound"
 	CONDITION_MESSAGE_RESOURCE_FOUND = "Resource found"
 )
 
 func (reconcile *ConfigReconciler) setConditionResourceFound(ctx context.Context, config *hubofhubsv1alpha1.Config) error {
 	if !reconcile.containsCondition(ctx, config, CONDITION_REASON_RESOURCE_FOUND) {
-		return utils.AppendCondition(ctx, reconcile.Client, config, CONDITION_TYPE_RESOURCE_FOUND,
+		return AppendCondition(ctx, reconcile.Client, config, CONDITION_TYPE_RESOURCE_FOUND,
 			CONDITION_STATUS_TRUE, CONDITION_REASON_RESOURCE_FOUND, CONDITION_MESSAGE_RESOURCE_FOUND)
 	}
 	return nil
 }
 
-// NOTE: the status of DATABASE_EXISTS can be True or False
+// NOTE: the status of DatabaseInitialized can be True or False
 const (
 	CONDITION_TYPE_DATABASE_INIT    = "DatabaseInitialized"
 	CONDITION_REASON_DATABASE_INIT  = "DatabaseInitialized"
@@ -41,13 +46,13 @@ const (
 
 func (reconcile *ConfigReconciler) setConditionDatabaseInit(ctx context.Context, config *hubofhubsv1alpha1.Config, status metav1.ConditionStatus) error {
 	if !reconcile.containsCondition(ctx, config, CONDITION_REASON_DATABASE_INIT) {
-		return utils.AppendCondition(ctx, reconcile.Client, config, CONDITION_TYPE_DATABASE_INIT,
+		return AppendCondition(ctx, reconcile.Client, config, CONDITION_TYPE_DATABASE_INIT,
 			status, CONDITION_REASON_DATABASE_INIT, CONDITION_MESSAGE_DATABASE_INIT)
 	} else {
 		currentStatus := reconcile.getConditionStatus(ctx, config, CONDITION_TYPE_DATABASE_INIT)
 		if currentStatus != status {
 			reconcile.deleteCondition(ctx, config, CONDITION_TYPE_DATABASE_INIT, CONDITION_REASON_DATABASE_INIT)
-			return utils.AppendCondition(ctx, reconcile.Client, config, CONDITION_TYPE_DATABASE_INIT, status,
+			return AppendCondition(ctx, reconcile.Client, config, CONDITION_TYPE_DATABASE_INIT, status,
 				CONDITION_REASON_DATABASE_INIT, CONDITION_MESSAGE_DATABASE_INIT)
 		}
 	}
@@ -103,4 +108,21 @@ func (reconcile *ConfigReconciler) containConditionStatus(ctx context.Context,
 		}
 	}
 	return output
+}
+
+func AppendCondition(ctx context.Context, reconcilerClient client.Client, object client.Object, typeName string,
+	stats metav1.ConditionStatus, reason string, message string,
+) error {
+	conditions, ok := (object).(ICondition)
+	if ok {
+		condition := metav1.Condition{Type: typeName, Status: stats, Reason: reason, Message: message, LastTransitionTime: metav1.Time{Time: time.Now()}}
+		conditions.SetConditions(append(conditions.GetConditions(), condition))
+		err := reconcilerClient.Status().Update(ctx, object)
+		if err != nil {
+			return fmt.Errorf("custom status condition update failed: %v", err)
+		}
+	} else {
+		return fmt.Errorf("status condition cannot be set, resource does not support Conditions")
+	}
+	return nil
 }
